Fixed synchronization problems on the course generation.
"no free events for message" caused missing cells on the clients. It has been fixed by increasing the Max Sent Message Queue Size, waiting a bit between cell generations to give time to dispatch the generation messages and waiting a bit before starting the generation so the position of the generator can be sync on the clients.